#1134 PR review updates - AbstractExecutionPlansController removed, smaller updates
